refactor: remove duplicates
